+++ common/config/src/main/java/io/helidon/common/config/spi/ConfigProvider.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, 2023 Oracle and/or its affiliates.
+ * Copyright (c) 2022, 2025 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,17 +16,19 @@
 
 package io.helidon.common.config.spi;
 
-import io.helidon.common.config.Config;
-
 /**
  * Service loader provider interface to discover config implementation that would be used to
  * obtain a default configuration instance.
+ *
+ * @deprecated there is no replacement for this type, use factory methods on the implementation directly
  */
+@SuppressWarnings("removal")
+@Deprecated(forRemoval = true, since = "4.3.0")
 public interface ConfigProvider {
     /**
      * Create the default configuration instance.
      *
      * @return a new configuration
      */
-    Config create();
+    io.helidon.common.config.Config create();
 }