+++ common/socket/src/main/java/module-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, 2024 Oracle and/or its affiliates.
+ * Copyright (c) 2022, 2025 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,7 +22,7 @@ module io.helidon.common.socket {
     requires io.helidon.common;
     // even though this is part of API, it is not transitive,
     // as unless your module uses config, the API is not useful
-    requires io.helidon.common.config;
+    requires io.helidon.config;
 
     requires transitive io.helidon.common.buffers;
     requires transitive io.helidon.builder.api;