+++ data/sql/datasource/hikari/src/main/java/module-info.java
@@ -19,7 +19,7 @@
  */
 module io.helidon.data.sql.datasource.hikari {
 
-    requires io.helidon.common.config;
+    requires io.helidon.config;
     requires io.helidon.builder.api;
     requires io.helidon.service.registry;
     requires transitive io.helidon.data.sql.common;