+++ integrations/langchain4j/providers/jlama/src/main/java/module-info.java
@@ -33,7 +33,7 @@ module io.helidon.integrations.langchain4j.providers.jlama {
 
     requires transitive io.helidon.service.registry;
     requires transitive io.helidon.integrations.langchain4j;
-    requires transitive io.helidon.common.config;
+    requires transitive io.helidon.config;
     requires jlama.core;
     requires io.helidon.common;
     requires io.helidon.builder.api;